VideoDownloaderPro/requirements.txt
yt-dlp>=2023.12.30
Pillow>=10.0.0
requests>=2.31.0 